docs: add header and feedback link for playground page
