+++ .github/feedback_template.md
@@ -1,0 +1,22 @@
+---
+name: Feedback
+about: Leave feedback about Paragon and suggest an addition or improvement
+---
+
+## Feedback summary
+
+<!--
+Leave general feedback about Paragon here ↓
+-->
+
+## Improvements
+
+<!--
+Make a suggestion on how we can improve Paragon's documentation or components.
+-->
+
+## Bugs
+
+<!--
+Add a bug description if you found one.
+-->
