+++ www/package.json
@@ -25,7 +25,6 @@
     "lodash.debounce": "^4.0.8",
     "prism-react-renderer": "^1.3.3",
     "prop-types": "^15.8.1",
-    "query-string": "^8.1.0",
     "react": "^17.0.2",
     "react-docgen": "^5.4.2",
     "react-dom": "^17.0.2",
@@ -36,7 +35,7 @@
     "rehype-autolink-headings": "^5.1.0",
     "rehype-slug": "^4.0.1",
     "sass": "^1.53.0",
-    "sass-loader": "^13.2.2",
+    "sass-loader": "12.6.0",
     "uuid": "^9.0.0"
   },
   "keywords": [
