+++ www/playroom/FrameComponent.jsx
@@ -1,23 +1,28 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
-import { openedx, edxOrg } from './themes';
 
 export default function FrameComponent({ theme, children }) {
   return (
-    <main>
+    <main style={{ padding: '16px' }}>
       <Helmet>
         <link
           href="/static/openedx-theme.css"
-          rel={`stylesheet${theme === openedx ? '' : ' alternate'}`}
+          rel={`stylesheet${theme === 'openedx' ? '' : ' alternate'}`}
           type="text/css"
         />
         <link
           href="/static/edxorg-theme.css"
-          rel={`stylesheet${theme === edxOrg ? '' : ' alternate'}`}
+          rel={`stylesheet${theme === 'edX.org' ? '' : ' alternate'}`}
           type="text/css"
         />
       </Helmet>
       {children}
     </main>
   );
+}
+
+FrameComponent.propTypes = {
+  theme: PropTypes.oneOf(['openedx', 'edX.org']).isRequired,
+  children: PropTypes.node.isRequired,
 };
