+++ www/src/components/PageLayout.tsx
@@ -19,7 +19,7 @@ import {
   breakpoints,
 } from '~paragon-react';
 import ComponentsList from './ComponentsList';
-import Header from './Header';
+import Header from './header';
 import Menu from './Menu';
 import Settings from './Settings';
 import Toc from './Toc';
