+++ www/src/components/header/index.js
@@ -1,0 +1,2 @@
+export { default } from './Header';
+export { default as SiteTitle } from './SiteTitle';
