+++ www/src/scss/base.scss
@@ -4,7 +4,7 @@
 @import "../components/doc-elements";
 @import "../components/Settings";
 @import "../components/Toc";
-@import "../components/Header";
+@import "../components/header/Header";
 @import "../components/Search/Search";
 @import "../components/IconsTable";
 @import "../components/exampleComponents/ExamplePropsForm";
